Badge Manager Updates
- Badges added to game
- BadgeManager saves and loads states
- Minigames complete are added to the save/load slot
- Minigames can either call BadgeManager or SaveLoadManager to set that they're completed. This depends on wether minigame is loaded additively or in single

+++ Assets/GameWorld/Scripts/Minigame Hookups/WinMinigame.cs
@@ -20,7 +20,7 @@ namespace Methodyca.Minigames.SortGame
 
         public void CompleteMinigame()
         {
-            BadgeManager.instance.SetMinigameComplete(minigameID);
+            BadgeManager.instance.SetMinigameComplete((int)minigameID);
             gameHub.EndGame();
         }
     }
